Changed createParseTree function, changed add function

+++ SeventhHomework26.10/ParseTree/parseTree.c
@@ -6,7 +6,7 @@
 
 typedef struct TreeNode
 {
-    char value;
+    char* value;
     struct TreeNode* leftSon;
     struct TreeNode* rightSon;
     struct TreeNode* parent;
@@ -23,22 +23,22 @@ int operation(const int value1, const int value2,
     *errorCode = 0;
     switch (operation)
     {
-    case (int) '+':
-        return value1 + value2;
-    case (int) '-':
-        return value1 - value2;
-    case (int) '*':
-        return value1 * value2;
-    case (int) '/':
-        if (value2 == 0)
-        {
-            *errorCode = -1;
+        case (int) '+':
+            return value1 + value2;
+        case (int) '-':
+            return value1 - value2;
+        case (int) '*':
+            return value1 * value2;
+        case (int) '/':
+            if (value2 == 0)
+            {
+                *errorCode = -1;
+                return 0;
+            }
+            return value1 / value2;
+        default:
+            *errorCode = -2;
             return 0;
-        }
-        return value1 / value2;
-    default:
-        *errorCode = -2;
-        return 0;
     }
 }
 
@@ -54,51 +54,6 @@ bool inLine(const char string[], const char symbol)
     return false;
 }
 
-int add(ParseTree** tree, ParseTree* root, char value)
-{
-    TreeNode* newNode = calloc(1, sizeof(TreeNode));
-    if (newNode == NULL)
-    {
-        return -1;
-    }
-    bool isDigit = value >= '0' && value <= '9';
-    newNode->value = value;
-    if ((*tree)->root == NULL)
-    {
-        root->root = newNode;
-        (*tree)->root = newNode;
-        return 0;
-    }
-    if ((*tree)->root->leftSon == NULL)
-    {
-        newNode->parent = (*tree)->root;
-        (*tree)->root->leftSon = newNode;
-        if (!isDigit)
-        {
-            (*tree)->root = (*tree)->root->leftSon;
-        }
-        return 0;
-    }
-    if ((*tree)->root->rightSon == NULL)
-    {
-        newNode->parent = (*tree)->root;
-        (*tree)->root->rightSon = newNode;
-        if (!isDigit)
-        {
-            (*tree)->root = (*tree)->root->rightSon;
-        }
-        else
-        {
-            while((*tree)->root != root->root && (*tree)->root->rightSon != NULL)
-            {
-                (*tree)->root = (*tree)->root->parent;
-            }
-        }
-        return 0;
-    }
-    return -2;
-}
-
 void deleteTreeRecursive(TreeNode* root)
 {
     if (root == NULL)
@@ -106,14 +61,67 @@ void deleteTreeRecursive(TreeNode* root)
         return;
     }
     deleteTreeRecursive(root->leftSon);
+    free(root->leftSon);
     deleteTreeRecursive(root->rightSon);
-    free(root);
+    free(root->rightSon);
 }
 
 void deleteTree(ParseTree** root)
 {
+    if ((*root) == NULL)
+    {
+        return;
+    }
     deleteTreeRecursive((*root)->root);
-    *root = NULL;
+    free((*root)->root);
+    free(*root);
+    (*root) = NULL;
+}
+
+int add(ParseTree* root, TreeNode** currentNode, char* value)
+{
+    TreeNode* newNode = calloc(1, sizeof(TreeNode));
+    if (newNode == NULL)
+    {
+        return -1;
+    }
+    bool isDigit = value[0] >= '0' && value[0] <= '9'
+            || value[1] >= '0' && value[1] <= '9';
+    newNode->value = value;
+    if (*currentNode == NULL)
+    {
+        root->root = newNode;
+        *currentNode = newNode;
+        return 0;
+    }
+    if ((*currentNode)->leftSon == NULL)
+    {
+        newNode->parent = *currentNode;
+        (*currentNode)->leftSon = newNode;
+        if (!isDigit)
+        {
+            *currentNode = (*currentNode)->leftSon;
+        }
+        return 0;
+    }
+    if ((*currentNode)->rightSon == NULL)
+    {
+        newNode->parent = *currentNode;
+        (*currentNode)->rightSon = newNode;
+        if (!isDigit)
+        {
+            *currentNode = (*currentNode)->rightSon;
+        }
+        else
+        {
+            while(*currentNode != root->root && (*currentNode)->rightSon != NULL)
+            {
+                *currentNode = (*currentNode)->parent;
+            }
+        }
+        return 0;
+    }
+    return -2;
 }
 
 ParseTree* createTree(const char expression[], int *errorCode)
@@ -125,26 +133,82 @@ ParseTree* createTree(const char expression[], int *errorCode)
         *errorCode = -1;
         return NULL;
     }
-    ParseTree* currentNode = calloc(1, sizeof(ParseTree));
+    TreeNode** currentNode = calloc(1, sizeof(TreeNode*));
     if (currentNode == NULL)
     {
         *errorCode = -1;
         free(root);
         return NULL;
     }
-    const char operations[5] = { '+', '-', '*', '/' };
-    const unsigned long lengthOfExpression = strlen(expression);
-    for (unsigned long i = 0; i < lengthOfExpression; ++i)
+    const char* operations = "*/+-";
+    const int lengthOfExpression = strlen(expression);
+    int state = 0;
+    char* currentValue = NULL;
+    int j = 0;
+    for (int i = 0; i < lengthOfExpression; ++i)
     {
-        if (inLine(operations, expression[i]) || expression[i] >= '0' && expression[i] <= '9')
+        char c = expression[i];
+        switch (state)
         {
-            *errorCode = add(&currentNode, root, expression[i]);
-            if (*errorCode != 0)
+            case 0:
             {
-                deleteTree(&root);
-                free(root);
-                free(currentNode);
-                return NULL;
+                if (c == '-')
+                {
+                    state = 1;
+                    currentValue = calloc(15, sizeof(char));
+                    currentValue[j] = c;
+                    ++j;
+                    break;
+                }
+                if (inLine(operations, c))
+                {
+                    state = 0;
+                    currentValue = calloc(2, sizeof(char));
+                    currentValue[j] = c;
+                    j = 0;
+                    add(root, currentNode, currentValue);
+                    currentValue = NULL;
+                    break;
+                }
+                if (c >= '0' && c <= '9')
+                {
+                    state = 2;
+                    currentValue = calloc(15, sizeof(char));
+                    currentValue[j] = c;
+                    ++j;
+                    break;
+                }
+                break;
+            }
+            case 1:
+            {
+                if (c >= '0' && c <= '9')
+                {
+                    state = 2;
+                    currentValue[j] = c;
+                    ++j;
+                    break;
+                }
+                state = 0;
+                j = 0;
+                add(root, currentNode, currentValue);
+                currentValue = NULL;
+                break;
+            }
+            case 2:
+            {
+                if (c >= '0' && c <= '9')
+                {
+                    state = 2;
+                    currentValue[j] = c;
+                    ++j;
+                    break;
+                }
+                state = 0;
+                j = 0;
+                add(root, currentNode, currentValue);
+                currentValue = NULL;
+                break;
             }
         }
     }
@@ -154,25 +218,27 @@ ParseTree* createTree(const char expression[], int *errorCode)
 
 int calculateValueOfTree(ParseTree* root)
 {
-    if (root->root->value < '0' || root->root->value > '9')
+    if (root->root->value > '0' && root->root->value < '9')
     {
-        ParseTree* leftTree = calloc(1, sizeof(ParseTree));
-        ParseTree* rightTree = calloc(1, sizeof(ParseTree));
-        leftTree->root = root->root->leftSon;
-        rightTree->root = root->root->rightSon;
-        int leftValue = calculateValueOfTree(leftTree);
-        int rightValue = calculateValueOfTree(rightTree);
-        int errorCode = 0;
-        int result = operation(leftValue, rightValue, root->root->value, &errorCode);
-        if (errorCode != 0)
-        {
-            return -1;
-        }
+        return (root->root->value) - '0';
+    }
+    ParseTree* leftTree = calloc(1, sizeof(ParseTree));
+    ParseTree* rightTree = calloc(1, sizeof(ParseTree));
+    leftTree->root = root->root->leftSon;
+    rightTree->root = root->root->rightSon;
+    int leftValue = calculateValueOfTree(leftTree);
+    int rightValue = calculateValueOfTree(rightTree);
+    int errorCode = 0;
+    int result = operation(leftValue, rightValue, root->root->value, &errorCode);
+    if (errorCode != 0)
+    {
         free(leftTree);
         free(rightTree);
-        return result;
+        return -1;
     }
-    return (root->root->value) - '0';
+    free(leftTree);
+    free(rightTree);
+    return result;
 }
 
 void printTree(ParseTree* root)
